Fix markdown rendering: use lightweight renderer, move styles to shared
- Use a local markdown-it instance (no KaTeX/LaTeX) for instruction text
  in history view, avoiding the broken math rendering from missing KaTeX CSS
- Move markdownStyles to @shared/styles/ so both progress and settings views
  can import it without cross-webview coupling
- Add syntax highlighting via shared hljs for code blocks in instructions
- Settings view bundle reduced from ~7.6MB to ~4.2MB by not pulling in KaTeX
